+++ modules/extensions-support/src/main/java/org/mule/runtime/module/extension/internal/runtime/source/SourceAdapter.java
@@ -401,7 +401,6 @@ public class SourceAdapter implements Lifecycle, Restartable {
           .withExpressionManager(expressionManager)
           .withConfig(configurationInstance)
           .withProperty(COMPONENT_DECORATOR_FACTORY_KEY, cursorDecoratorFactory.componentDecoratorFactory(component))
-          .resolveCursors(false)
           .build();
     }
   }
